# common-platform-org-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-org-mgmt
  description: Organization Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/organization'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/banks/{bankId}/branches/{branchId}':
    get:
      tags:
        - Bank Branch Management
      summary: Get branch by ID for a bank
      description: Returns a branch of a specific bank based on its ID
      operationId: getBranchByIdForBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: branchId
          in: path
          description: ID of the branch to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '404':
          description: Bank or branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
    put:
      tags:
        - Bank Branch Management
      summary: Update branch for a bank
      description: Updates an existing branch of a specific bank with the provided details
      operationId: updateBranchForBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: branchId
          in: path
          description: ID of the branch to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDTO'
        required: true
      responses:
        '200':
          description: Branch successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '400':
          description: Invalid branch data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '404':
          description: Bank or branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
    delete:
      tags:
        - Bank Branch Management
      summary: Delete branch for a bank
      description: Deletes a branch of a specific bank based on its ID
      operationId: deleteBranchForBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: branchId
          in: path
          description: ID of the branch to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Branch successfully deleted
        '404':
          description: Bank or branch not found
        '500':
          description: Internal server error
  '/api/v1/banks/{bankId}':
    get:
      tags:
        - Bank Management
      summary: Get bank by ID
      description: Returns a bank based on its ID
      operationId: getBankById
      parameters:
        - name: bankId
          in: path
          description: ID of the bank to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved bank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
    put:
      tags:
        - Bank Management
      summary: Update bank
      description: Updates an existing bank with the provided details
      operationId: updateBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankDTO'
        required: true
      responses:
        '200':
          description: Bank successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '400':
          description: Invalid bank data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
    delete:
      tags:
        - Bank Management
      summary: Delete bank
      description: Deletes a bank based on its ID
      operationId: deleteBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Bank successfully deleted
        '404':
          description: Bank not found
        '500':
          description: Internal server error
...
  '/api/v1/banks/{bankId}/branches/filter':
    post:
      tags:
        - Bank Branch Management
      summary: Get all branches for a bank with filtering
      description: >-
        Returns a paginated list of branches for a specific bank based on filter
        criteria
      operationId: filterBranchesForBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBranchDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved branches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchDTO'
  '/api/v1/banks/{bankId}/branches':
    post:
      tags:
        - Bank Branch Management
      summary: Create a new branch for a bank
      description: Creates a new branch for a specific bank with the provided details
      operationId: createBranchForBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDTO'
        required: true
      responses:
        '201':
          description: Branch successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '400':
          description: Invalid branch data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
  /api/v1/banks/filter:
    post:
      tags:
        - Bank Management
      summary: Get all banks with filtering
      description: Returns a paginated list of banks based on filter criteria
      operationId: filterBanks
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBankDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved banks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankDTO'
...
components:
  schemas:
...
    BranchDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bankId:
          type: integer
          format: int64
        regionId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        addressLine:
          type: string
        postalCode:
          type: string
        city:
          type: string
        state:
          type: string
        countryId:
          type: integer
          format: int64
        timeZoneId:
          type: integer
          format: int64
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        isActive:
          type: boolean
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    BankDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        websiteUrl:
          type: string
        addressLine:
          type: string
        postalCode:
          type: string
        city:
          type: string
        state:
          type: string
        countryId:
          type: integer
          format: int64
        timeZoneId:
          type: integer
          format: int64
        isActive:
          type: boolean
        establishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    FilterRequestBranchDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BranchDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for branches
    PaginationResponseBranchDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BranchDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    FilterRequestBankDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BankDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for banks
    PaginationResponseBankDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BankDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.